requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2
django-cors-headers==3.14.0
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
psycopg2-binary==2.9.7
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
typing-extensions==4.7.1
urllib3==1.26.15
